Monitor removal of files, too
If a file disappears (deleted or moved to a different volume), pop
up an alert to either ignore this or call the OpenFileDialog to
diff two new files.

When calling the OpenFileDialog, add our calling window to the BMessege
and have it quit our window _after_ showing the OpenFileDialog. Otherwise
the whole app will quit when there are no windows left around...

Updated en.catkeys.