Open OpenFilesDialog at the calling window's position
Instead of using the last saved windows position from the settings,
add the calling window's position as a BRect in the message from
the window to the app.
In case the call to show the OpenFilesDialog does not come from a
window, but from a launch of the app without files to diff, use
the loaded settings window_frame.