Accept two dropped files, too
If more than 1 valid file is dropped on the window, don't mind over which
pane they were dropped. Open the first on the left, the second on the right.

Accepting more than two files doen't make sense, because the order in which
they are dropped is not defined, and only by luck you intend to compare
files alphabetically by file name...